Rename: 파일 폴더 변경
폴더 생성 후 파일 위치 변경
*InputSystem.inputCars return 값 형식 변경
