+++ src/main/java/racingcar/Utils/GameConstants.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.Utils;
 
 public class GameConstants {
     public final static String INPUT_CAR_NAME_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
